Move workspace functionality into partials.
