+++ src/DynamoCore/DynamoCore.csproj
@@ -117,8 +117,12 @@ limitations under the License.
     <Compile Include="Configuration\Context.cs" />
     <Compile Include="Core\DynamoMigrator.cs" />
     <Compile Include="Exceptions\LibraryLoadFailedException.cs" />
+    <Compile Include="Graph\Workspaces\Layout.cs" />
+    <Compile Include="Graph\Workspaces\NodeToCodeReconnection.cs" />
+    <Compile Include="Graph\Workspaces\Presets.cs" />
     <Compile Include="Graph\Workspaces\Serialiaztion\SerializationConverters.cs" />
     <Compile Include="Graph\Workspaces\Serialiaztion\Workspaces.cs" />
+    <Compile Include="Graph\Workspaces\UndoRedo.cs" />
     <Compile Include="Logging\AnalyticsService.cs" />
     <Compile Include="Logging\DynamoAnalyticsClient.cs" />
     <Compile Include="Logging\IAnalyticsSession.cs" />
