+++ src/DynamoCore/Graph/Workspaces/WorkspaceModel.cs
@@ -33,11 +33,8 @@ namespace Dynamo.Graph.Workspaces
     /// Represents base class for all kind of workspaces which contains general data
     /// such as Name, collections of nodes, notes, annotations, etc.
     /// </summary>
-    public abstract class WorkspaceModel : NotificationObject, ILocatable, IUndoRedoRecorderClient, ILogSource, IDisposable, IWorkspaceModel
+    public abstract partial class WorkspaceModel : NotificationObject, ILocatable, IUndoRedoRecorderClient, ILogSource, IDisposable, IWorkspaceModel
     {
-
-       
-
         #region private/internal members
 
         /// <summary>
@@ -1123,408 +1120,6 @@ namespace Dynamo.Graph.Workspaces
         }
 
         /// <summary>
-        /// This function wraps a few methods on the workspace model layer
-        /// to set up and run the graph layout algorithm.
-        /// </summary>
-        internal void DoGraphAutoLayout()
-        {
-            if (Nodes.Count() < 2) return;
-
-            var selection = DynamoSelection.Instance.Selection;
-
-            // Check if all the selected models are groups
-            bool isGroupLayout = selection.Count > 0 &&
-                selection.All(x => x is AnnotationModel ||
-                    selection.OfType<AnnotationModel>().Any(g => g.Nodes.Contains(x)));
-
-            GenerateCombinedGraph(isGroupLayout);
-            RecordUndoGraphLayout(isGroupLayout);
-
-            // Generate subgraphs separately for each cluster
-            SubgraphClusters.ForEach(
-                x => GenerateSeparateSubgraphs(new HashSet<GraphLayout.Node>(x)));
-
-            // Deselect all nodes
-            SubgraphClusters.ForEach(c => c.ForEach(x => x.IsSelected = false));
-
-            // Run layout algorithm for each subgraph
-            LayoutSubgraphs.Skip(1).ToList().ForEach(g => RunLayoutSubgraph(g, isGroupLayout));
-            AvoidSubgraphOverlap();
-            SaveLayoutGraph();
-
-            // Restore the workspace model selection information
-            selection.ToList().ForEach(x => x.Select());
-        }
-
-        /// <summary>
-        /// This method extracts all models from the workspace and puts them
-        /// into the combined graph object, LayoutSubgraphs.First()
-        /// <param name="isGroupLayout">True if all the selected models are groups.</param>
-        /// </summary>
-        private void GenerateCombinedGraph(bool isGroupLayout)
-        {
-            LayoutSubgraphs = new List<GraphLayout.Graph>();
-            LayoutSubgraphs.Add(new GraphLayout.Graph());
-
-            GraphLayout.Graph combinedGraph = LayoutSubgraphs.First();
-            SubgraphClusters = new List<List<GraphLayout.Node>>();
-
-            if (!isGroupLayout)
-            {
-                foreach (AnnotationModel group in Annotations)
-                {
-                    // Treat a group as a graph layout node/vertex
-                    combinedGraph.AddNode(group.GUID, group.Width, group.Height, group.X, group.Y,
-                        group.IsSelected || DynamoSelection.Instance.Selection.Count == 0);
-                }
-            }
-
-            foreach (NodeModel node in Nodes)
-            {
-                if (!isGroupLayout)
-                {
-                    AnnotationModel group = Annotations.Where(
-                        g => g.Nodes.Contains(node)).ToList().FirstOrDefault();
-
-                    // Do not process nodes within groups
-                    if (group == null)
-                    {
-                        combinedGraph.AddNode(node.GUID, node.Width, node.Height, node.X, node.Y,
-                            node.IsSelected || DynamoSelection.Instance.Selection.Count == 0);
-                    }
-                }
-                else
-                {
-                    // Process all nodes inside the selection
-                    combinedGraph.AddNode(node.GUID, node.Width, node.Height, node.X, node.Y,
-                        node.IsSelected || DynamoSelection.Instance.Selection.Count == 0);
-                }
-            }
-
-            foreach (ConnectorModel edge in Connectors)
-            {
-                if (!isGroupLayout)
-                {
-                    AnnotationModel startGroup = null, endGroup = null;
-                    startGroup = Annotations.Where(
-                        g => g.Nodes.Contains(edge.Start.Owner)).ToList().FirstOrDefault();
-                    endGroup = Annotations.Where(
-                        g => g.Nodes.Contains(edge.End.Owner)).ToList().FirstOrDefault();
-
-                    // Treat a group as a node, but do not process edges within a group
-                    if (startGroup == null || endGroup == null || startGroup != endGroup)
-                    {
-                        combinedGraph.AddEdge(
-                            startGroup == null ? edge.Start.Owner.GUID : startGroup.GUID,
-                            endGroup == null ? edge.End.Owner.GUID : endGroup.GUID,
-                            edge.Start.Center.X, edge.Start.Center.Y, edge.End.Center.X, edge.End.Center.Y);
-                    }
-                }
-                else
-                {
-                    // Edges within a group are also processed
-                    combinedGraph.AddEdge(edge.Start.Owner.GUID, edge.End.Owner.GUID,
-                        edge.Start.Center.X, edge.Start.Center.Y, edge.End.Center.X, edge.End.Center.Y);
-                }
-            }
-
-            foreach (NoteModel note in Notes)
-            {
-                AnnotationModel group = Annotations.Where(
-                    g => g.Nodes.Contains(note)).ToList().FirstOrDefault();
-
-                GraphLayout.Node nd = null;
-
-                if (!isGroupLayout || group == null)
-                {
-                    // If note is not part of a group, link to the nearest node in the graph
-                    nd = combinedGraph.Nodes.OrderBy(node =>
-                        Math.Pow(node.X + node.Width / 2 - note.X - note.Width / 2, 2) +
-                        Math.Pow(node.Y + node.Height / 2 - note.Y - note.Height / 2, 2)).FirstOrDefault();
-                }
-                else
-                {
-                    // If note is part of a group, link to the nearest node in the group
-                    NodeModel ndm = group.Nodes.OfType<NodeModel>().OrderBy(node =>
-                        Math.Pow(node.X + node.Width / 2 - note.X - note.Width / 2, 2) +
-                        Math.Pow(node.Y + node.Height / 2 - note.Y - note.Height / 2, 2)).FirstOrDefault();
-
-                    // Skip processing the group if there is no node in the group
-                    if (ndm == null) continue;
-
-                    // If the nearest point is a node model
-                    nd = combinedGraph.FindNode(ndm.GUID);
-
-                    // If the nearest point is a group model
-                    nd = nd ?? combinedGraph.FindNode(group.GUID);
-                }
-
-                // Otherwise, leave the note unchanged
-                if (nd != null)
-                {
-                    nd.LinkNote(note, note.Width, note.Height);
-                }
-            }
-
-            if (!isGroupLayout)
-            {
-                // Add all nodes to one big cluster
-                List<GraphLayout.Node> bigcluster = new List<GraphLayout.Node>();
-                bigcluster.AddRange(combinedGraph.Nodes);
-                SubgraphClusters.Add(bigcluster);
-            }
-            else
-            {
-                // Each group becomes one cluster
-                foreach (AnnotationModel group in DynamoSelection.Instance.Selection.OfType<AnnotationModel>())
-                {
-                    List<GraphLayout.Node> cluster = new List<GraphLayout.Node>();
-                    cluster.AddRange(group.Nodes.OfType<NodeModel>().Select(x => combinedGraph.FindNode(x.GUID)));
-                    SubgraphClusters.Add(cluster);
-                }
-            }
-
-        }
-
-        /// <summary>
-        /// This method adds relevant models to the undo recorder.
-        /// </summary>
-        /// <param name="isGroupLayout">True if all the selected models are groups.</param>
-        private void RecordUndoGraphLayout(bool isGroupLayout)
-        {
-            List<ModelBase> undoItems = new List<ModelBase>();
-
-            if (!isGroupLayout)
-            {
-                // Add all selected items to the undo recorder
-                undoItems.AddRange(Nodes);
-                undoItems.AddRange(Notes);
-                if (DynamoSelection.Instance.Selection.Count > 0)
-                {
-                    undoItems = undoItems.Where(x => x.IsSelected).ToList();
-                }
-            }
-            else
-            {
-                // Add all models inside selected groups
-                foreach (var group in Annotations)
-                {
-                    if (group.IsSelected)
-                    {
-                        group.Nodes.OfType<NodeModel>().ToList().ForEach(x => x.IsSelected = false);
-                        undoItems.AddRange(group.Nodes);
-                    }
-                }
-            }
-
-            WorkspaceModel.RecordModelsForModification(undoItems, UndoRecorder);
-        }
-
-        /// <summary>
-        /// This method repeatedly takes a selected node in the combined graph and
-        /// uses breadth-first search to find all other nodes in the same subgraph
-        /// until all selected nodes have been processed.
-        /// </summary>
-        /// <param name="nodes">A cluster of nodes to be separated into subgraphs.</param>
-        private void GenerateSeparateSubgraphs(HashSet<GraphLayout.Node> nodes)
-        {
-            int processed = 0;
-            var combinedGraph = LayoutSubgraphs.First();
-            GraphLayout.Graph graph = new GraphLayout.Graph();
-            Queue<GraphLayout.Node> queue = new Queue<GraphLayout.Node>();
-
-            while (nodes.Count(n => n.IsSelected) > 0)
-            {
-                GraphLayout.Node currentNode;
-
-                if (queue.Count == 0)
-                {
-                    if (graph.Nodes.Count > 0)
-                    {
-                        // Save the subgraph and subtract these nodes from the combined graph
-
-                        LayoutSubgraphs.Add(graph);
-                        nodes.ExceptWith(graph.Nodes);
-                        combinedGraph.Nodes.ExceptWith(graph.Nodes);
-                        graph = new GraphLayout.Graph();
-                    }
-                    if (nodes.Count(n => n.IsSelected) == 0) break;
-
-                    currentNode = nodes.FirstOrDefault(n => n.IsSelected);
-                    graph.Nodes.Add(currentNode);
-                }
-                else
-                {
-                    currentNode = queue.Dequeue();
-                }
-
-                // Find all nodes in the selection which are connected directly
-                // to the left or to the right to the currentNode
-
-                var selectedNodes = currentNode.RightEdges.Select(e => e.EndNode)
-                    .Union(currentNode.LeftEdges.Select(e => e.StartNode))
-                    .Where(x => nodes.Contains(x) && x.IsSelected)
-                    .Except(graph.Nodes).ToList();
-                graph.Nodes.UnionWith(selectedNodes);
-                graph.Edges.UnionWith(currentNode.RightEdges);
-                graph.Edges.UnionWith(currentNode.LeftEdges);
-
-                // If any of the incident edges are connected to unselected (outside) nodes
-                // then mark these edges as anchors.
-
-                graph.AnchorRightEdges.UnionWith(currentNode.RightEdges.Where(e => !e.EndNode.IsSelected));
-                graph.AnchorLeftEdges.UnionWith(currentNode.LeftEdges.Where(e => !e.StartNode.IsSelected));
-
-                foreach (var node in selectedNodes)
-                {
-                    queue.Enqueue(node);
-                    processed++;
-                }
-            }
-        }
-
-        /// <summary>
-        /// This function calls the graph layout algorithm methods.
-        /// </summary>
-        /// <param name="graph">The subgraph to be processed.</param>
-        /// <param name="isGroupLayout">True if all selected models are groups.</param>
-        private void RunLayoutSubgraph(GraphLayout.Graph graph, bool isGroupLayout)
-        {
-            // Select relevant nodes
-            graph.Nodes.ToList().ForEach(x => x.IsSelected = true);
-
-            // Save subgraph position before running the layout
-            graph.RecordInitialPosition();
-
-            // Sugiyama algorithm steps
-            graph.RemoveCycles();
-            graph.AssignLayers();
-            graph.OrderNodes();
-
-            // Node and graph positioning
-            graph.DistributeNodePosition();
-            graph.SetGraphPosition(isGroupLayout);
-
-            // Reset layer information and deselect nodes
-            graph.ResetLayers();
-            graph.Nodes.ToList().ForEach(x => x.IsSelected = false);
-        }
-
-        /// <summary>
-        /// This method repeatedly shifts subgraphs away vertically from each other
-        /// when there are any two nodes from different subgraphs overlapping.
-        /// </summary>
-        private void AvoidSubgraphOverlap()
-        {
-            bool done;
-
-            do
-            {
-                done = true;
-
-                foreach (var g1 in LayoutSubgraphs.Skip(1))
-                {
-                    foreach (var g2 in LayoutSubgraphs.Skip(1))
-                    {
-                        // The first subgraph's center point must be higher than the second subgraph
-                        if (!g1.Equals(g2) && (g1.GraphCenterY + g1.OffsetY <= g2.GraphCenterY + g2.OffsetY))
-                        {
-                            var g1nodes = g1.Nodes.OrderBy(n => n.Y + n.TotalHeight);
-                            var g2nodes = g2.Nodes.OrderBy(n => n.Y);
-
-                            foreach (var node1 in g1nodes)
-                            {
-                                foreach (var node2 in g2nodes)
-                                {
-                                    // If any two nodes from these two different subgraphs overlap
-                                    if ((node1.Y + node1.TotalHeight + GraphLayout.Graph.VerticalNodeDistance + g1.OffsetY > node2.Y + g2.OffsetY) &&
-                                        (((node1.X <= node2.X) && (node1.X + node1.Width + GraphLayout.Graph.HorizontalNodeDistance > node2.X)) ||
-                                        ((node2.X <= node1.X) && (node2.X + node2.Width + GraphLayout.Graph.HorizontalNodeDistance > node1.X))))
-                                    {
-                                        // Shift the first subgraph to the top and the second subgraph to the bottom
-                                        g1.OffsetY -= 5;
-                                        g2.OffsetY += 5;
-                                        done = false;
-                                    }
-                                    if (!done) break;
-                                }
-                                if (!done) break;
-                            }
-                        }
-                    }
-                }
-            } while (!done);
-        }
-
-        /// <summary>
-        /// This method pushes changes from the GraphLayout.Graph objects
-        /// back to the workspace models.
-        /// </summary>
-        private void SaveLayoutGraph()
-        {
-            // Assign coordinates to nodes inside groups
-            foreach (var group in Annotations)
-            {
-                GraphLayout.Graph graph = LayoutSubgraphs
-                    .FirstOrDefault(g => g.FindNode(group.GUID) != null);
-
-                if (graph != null)
-                {
-                    GraphLayout.Node n = graph.FindNode(group.GUID);
-
-                    double deltaX = n.X - group.X;
-                    double deltaY = n.Y - group.Y + graph.OffsetY;
-                    foreach (var node in group.Nodes.OfType<NodeModel>())
-                    {
-                        node.X += deltaX;
-                        node.Y += deltaY;
-                        node.ReportPosition();
-                    }
-
-                    foreach (NoteModel note in n.LinkedNotes)
-                    {
-                        if (note.IsSelected || DynamoSelection.Instance.Selection.Count == 0)
-                        {
-                            note.X += deltaX;
-                            note.Y += deltaY;
-                            note.ReportPosition();
-                        }
-                    }
-                }
-            }
-
-            // Assign coordinates to nodes outside groups
-            foreach (var node in Nodes)
-            {
-                GraphLayout.Graph graph = LayoutSubgraphs
-                    .FirstOrDefault(g => g.FindNode(node.GUID) != null);
-
-                if (graph != null)
-                {
-                    GraphLayout.Node n = graph.FindNode(node.GUID);
-                    double offsetY = graph.OffsetY;
-
-                    node.X = n.X;
-                    node.Y = n.Y + n.NotesHeight + offsetY;
-                    node.ReportPosition();
-                    HasUnsavedChanges = true;
-
-                    double noteOffset = -n.NotesHeight;
-                    foreach (NoteModel note in n.LinkedNotes)
-                    {
-                        if (note.IsSelected || DynamoSelection.Instance.Selection.Count == 0)
-                        {
-                            note.X = node.X;
-                            note.Y = node.Y + noteOffset;
-                            noteOffset += note.Height + GraphLayout.Graph.VerticalNoteDistance;
-                            note.ReportPosition();
-                        }
-                    }
-                }
-            }
-        }
-
-        /// <summary>
         /// this sets the event on Annotation. This event return the model from the workspace.
         /// When a model is ungrouped from a group, that model will be deleted from that group.
         /// So, when UNDO execution, cannot get that model from that group, it has to get from the workspace.
@@ -1643,102 +1238,6 @@ namespace Dynamo.Graph.Workspaces
         internal void IncrementPasteOffset()
         {
             this.currentPasteOffset = (this.currentPasteOffset + PasteOffsetStep) % PasteOffsetMax;
-        }
-
-        #endregion
-
-        #region Presets
-
-        /// <summary>
-        ///  this method creates a new preset state from a set of NodeModels and adds this new state to this presets collection
-        /// </summary>
-        /// <param name="name">the name of preset state</param>
-        /// <param name="description">a description of what the state does</param>
-        /// <param name="currentSelection">a set of NodeModels that are to be serialized in this state</param>
-        private PresetModel AddPresetCore(string name, string description, IEnumerable<NodeModel> currentSelection)
-        {
-            if (currentSelection == null || currentSelection.Count() < 1)
-            {
-                throw new ArgumentException("currentSelection is empty or null");
-            }
-            var inputs = currentSelection;
-
-            var newstate = new PresetModel(name, description, inputs);
-            if (Presets.Any(x => x.GUID == newstate.GUID))
-            {
-                throw new ArgumentException("duplicate id in collection");
-            }
-
-            presets.Add(newstate);
-            return newstate;
-        }
-
-        /// <summary>
-        /// Removes a specified <see cref="PresetModel"/> object from the preset collection of the workspace.
-        /// </summary>
-        /// <param name="state"><see cref="PresetModel"/> object to remove.</param>
-        public void RemovePreset(PresetModel state)
-        {
-            if (Presets.Contains(state))
-            {
-                presets.Remove(state);
-            }
-        }
-
-        internal void ApplyPreset(PresetModel state)
-        {
-            if (state == null)
-            {
-                Log("Attempted to apply a PresetState that was null");
-                return;
-            }
-            //start an undoBeginGroup
-            using (var undoGroup = this.undoRecorder.BeginActionGroup())
-            {
-               //reload each node, and record each each modification in the undogroup
-                foreach (var node in state.Nodes)
-                {
-                    //check that node still exists in this workspace,
-                    //otherwise bail on this node, check by GUID instead of nodemodel
-                    if (nodes.Select(x=>x.GUID).Contains(node.GUID))
-                    {
-                        var originalpos = node.Position;
-                        var serializedNode = state.SerializedNodes.ToList().Find(x => Guid.Parse(x.GetAttribute("guid")) == node.GUID);
-                        //overwrite the xy coords of the serialized node with the current position, so the node is not moved
-                        serializedNode.SetAttribute("x", originalpos.X.ToString(CultureInfo.InvariantCulture));
-                        serializedNode.SetAttribute("y", originalpos.Y.ToString(CultureInfo.InvariantCulture));
-                        serializedNode.SetAttribute("isPinned", node.PreviewPinned.ToString());
-
-                        this.undoRecorder.RecordModificationForUndo(node);
-                        this.ReloadModel(serializedNode);
-                    }
-                }
-                //select all the modified nodes in the UI
-                DynamoSelection.Instance.ClearSelection();
-                foreach(var node in state.Nodes)
-                {
-                    DynamoSelection.Instance.Selection.Add(node);
-                }
-            }
-        }
-
-        internal PresetModel AddPreset(string name, string description, IEnumerable<Guid> IDSToSave)
-        {
-                //lookup the nodes by their ID, can also check that we find all of them....
-                var nodesFromIDs = this.Nodes.Where(node => IDSToSave.Contains(node.GUID));
-                //access the presetsCollection and add a new state based on the current selection
-                var newpreset = this.AddPresetCore(name, description, nodesFromIDs);
-                HasUnsavedChanges = true;
-                return newpreset;
-        }
-
-        /// <summary>
-        /// Adds a specified collection <see cref="PresetModel"/> objects to the preset collection of the workspace.
-        /// </summary>
-        /// <param name="presetCollection"><see cref="PresetModel"/> objects to add.</param>
-        public void ImportPresets(IEnumerable<PresetModel> presetCollection)
-        {
-            presets.AddRange(presetCollection);
         }
 
         #endregion
@@ -2072,570 +1571,6 @@ namespace Dynamo.Graph.Workspaces
             DynamoSelection.Instance.Selection.AddRange(codeBlockNodes);
 
             RequestRun();
-        }
-
-        #endregion
-
-        #region Undo/Redo Supporting Methods
-
-        internal void Undo()
-        {
-            if (null != undoRecorder)
-            {
-                undoRecorder.Undo();
-
-                // http://adsk-oss.myjetbrains.com/youtrack/issue/MAGN-7883
-                // Request run for every undo action
-                RequestRun();
-            }
-        }
-
-        internal void Redo()
-        {
-            if (null != undoRecorder)
-            {
-                undoRecorder.Redo();
-
-                // http://adsk-oss.myjetbrains.com/youtrack/issue/MAGN-7883
-                // Request run for every redo action
-                RequestRun();
-            }
-        }
-
-        internal void ClearUndoRecorder()
-        {
-            if (null != undoRecorder)
-                undoRecorder.Clear();
-        }
-
-        // See RecordModelsForModification below for more details.
-        internal static void RecordModelForModification(ModelBase model, UndoRedoRecorder recorder)
-        {
-            if (null != model)
-            {
-                var models = new List<ModelBase> { model };
-                RecordModelsForModification(models, recorder);
-            }
-        }
-
-        /// <summary>
-        /// TODO(Ben): This method is exposed this way for external codes (e.g.
-        /// the DragCanvas) to record models before they are modified. This is
-        /// by no means ideal. The ideal case of course is for ALL codes that
-        /// end up modifying models to be folded back into WorkspaceViewModel in
-        /// the form of commands. These commands then internally record those
-        /// affected models before updating them. We need this method to be gone
-        /// sooner than later.
-        /// </summary>
-        /// <param name="models">The models to be recorded for undo.</param>
-        /// <param name="recorder"></param>
-        internal static void RecordModelsForModification(List<ModelBase> models, UndoRedoRecorder recorder)
-        {
-            if (null == recorder)
-                return;
-            if (!ShouldProceedWithRecording(models))
-                return;
-
-            using (recorder.BeginActionGroup())
-            {
-                foreach (var model in models)
-                    recorder.RecordModificationForUndo(model);
-            }
-        }
-
-        internal static void RecordModelsForUndo(Dictionary<ModelBase, UndoRedoRecorder.UserAction> models, UndoRedoRecorder recorder)
-        {
-            if (null == recorder)
-                return;
-            if (!ShouldProceedWithRecording(models))
-                return;
-            
-            if (null != savedModels)
-            {
-                // Before an existing connector is reconnected, we have one action group
-                // which records the deletion of the connector. Pop that out so that we can
-                // record the deletion and reconnection in one action group.
-                recorder.PopFromUndoGroup();
-            }
-
-            using (recorder.BeginActionGroup())
-            {
-                if (null != savedModels)
-                {
-                    foreach (var modelPair in savedModels)
-                    {
-                        recorder.RecordDeletionForUndo(modelPair);
-                    }
-                    savedModels = null;
-                }
-                foreach (var modelPair in models)
-                {
-                    switch (modelPair.Value)
-                    {
-                        case UndoRedoRecorder.UserAction.Creation:
-                            recorder.RecordCreationForUndo(modelPair.Key);
-                            break;
-                        case UndoRedoRecorder.UserAction.Deletion:
-                            recorder.RecordDeletionForUndo(modelPair.Key);
-                            break;
-                        case UndoRedoRecorder.UserAction.Modification:
-                            recorder.RecordModificationForUndo(modelPair.Key);
-                            break;
-                    }
-                }
-            }
-        }
-
-        internal void RecordCreatedModel(ModelBase model)
-        {
-            if (null == model) return;
-
-            using (undoRecorder.BeginActionGroup())
-            {
-                undoRecorder.RecordCreationForUndo(model);
-            }
-        }
-
-        internal void RecordCreatedModels(List<ModelBase> models)
-        {
-            if (!ShouldProceedWithRecording(models))
-                return; // There's nothing created.
-
-            using (undoRecorder.BeginActionGroup())
-            {
-                foreach (ModelBase model in models)
-                    undoRecorder.RecordCreationForUndo(model);
-            }
-        }
-
-        internal void RecordAndDeleteModels(List<ModelBase> models)
-
-        {
-            if (!ShouldProceedWithRecording(models))
-                return; // There's nothing for deletion.
-
-            // Gather a list of connectors first before the nodes they connect
-            // to are deleted. We will have to delete the connectors first
-            // before
-
-            using (undoRecorder.BeginActionGroup()) // Start a new action group.
-            {
-                foreach (var model in models)
-                {
-                    if (model is NoteModel)
-                    {
-                        // Take a snapshot of the note before it goes away.
-                        undoRecorder.RecordDeletionForUndo(model);
-                        RemoveNote(model as NoteModel);
-                    }
-                    else if (model is AnnotationModel)
-                    {
-                        undoRecorder.RecordDeletionForUndo(model);
-                        RemoveAnnotation(model as AnnotationModel);
-                    }
-
-                    else if (model is PresetModel)
-                    {
-                        undoRecorder.RecordDeletionForUndo(model);
-                        RemovePreset(model as PresetModel);
-                    }
-
-                    else if (model is NodeModel)
-                    {
-                        // Just to make sure we don't end up deleting nodes from
-                        // another workspace (potentially two issues: the node was
-                        // having its "Workspace" pointing to another workspace,
-                        // or the selection set was not quite set up properly.
-                        //
-                        var node = model as NodeModel;
-                        Debug.Assert(Nodes.Contains(node));
-
-                        bool silentFlag = node.RaisesModificationEvents;
-                        node.RaisesModificationEvents = false;
-
-                        // Note that AllConnectors is duplicated as a separate list
-                        // by calling its "ToList" method. This is the because the
-                        // "Connectors.Remove" will modify "AllConnectors", causing
-                        // the Enumerator in this "foreach" to become invalid.
-                        foreach (var conn in node.AllConnectors.ToList())
-                        {
-                            conn.Delete();
-                            undoRecorder.RecordDeletionForUndo(conn);
-                        }
-
-                        node.RaisesModificationEvents = silentFlag;
-
-                        // Take a snapshot of the node before it goes away.
-                        undoRecorder.RecordDeletionForUndo(node);
-
-                        RemoveAndDisposeNode(node);
-                    }
-                    else if (model is ConnectorModel)
-                    {
-                        undoRecorder.RecordDeletionForUndo(model);
-                    }
-                }
-
-                RequestRun();
-
-            } // Conclude the deletion.
-        }
-
-        internal void DeleteSavedModels()
-        {
-            savedModels = null;
-        }
-
-        internal void SaveAndDeleteModels(List<ModelBase> models)
-        {
-            if (null != models)
-            {
-                // If an existing connector is grabbed (to be reconnected), save the 
-                // models for deletion later in one action group.
-                savedModels = models;
-
-                // After saving the models, delete them from the workspace
-                // in one action group.
-                RecordAndDeleteModels(models);
-            }
-        }
-
-        internal void RecordGroupModelBeforeUngroup(AnnotationModel annotation)
-        {
-            using (undoRecorder.BeginActionGroup()) // Start a new action group.
-            {
-                undoRecorder.RecordModificationForUndo(annotation);
-            }
-        }
-
-        private static bool ShouldProceedWithRecording(List<ModelBase> models)
-        {
-            if (null == models)
-                return false;
-
-            models.RemoveAll(x => x == null);
-            return models.Count > 0;
-        }
-
-        private static bool ShouldProceedWithRecording(
-            Dictionary<ModelBase, UndoRedoRecorder.UserAction> models)
-        {
-            return (null != models && (models.Count > 0));
-        }
-
-        #endregion
-
-        #region IUndoRedoRecorderClient Members
-
-        /// <summary>
-        /// Deletes <see cref="ModelBase"/> object given by <see cref="XmlElement"/>
-        /// from a corresponding collection of the workspace.
-        /// </summary>
-        /// <param name="modelData"><see cref="ModelBase"/> object given by <see cref="XmlElement"/></param>
-        public void DeleteModel(XmlElement modelData)
-        {
-            //When there is a Redo operation, model is removed from
-            //the workspace but the model is "not disposed" from memory.
-            //Identified this when redo operation is performed on groups
-            ModelBase model = GetModelForElement(modelData);
-
-            if (model is NoteModel)
-            {
-                var note = model as NoteModel;
-                RemoveNote(note);
-                note.Dispose();
-            }
-            else if (model is AnnotationModel)
-            {
-                RemoveGroup(model);
-            }
-            else if (model is PresetModel)
-            {
-                RemovePreset(model as PresetModel);
-            }
-            else if (model is ConnectorModel)
-            {
-                var connector = model as ConnectorModel;
-                connector.Delete();
-            }
-            else if (model is NodeModel)
-            {
-                RemoveAndDisposeNode(model as NodeModel);
-            }
-            else
-            {
-                // If it gets here we obviously need to handle it.
-                throw new InvalidOperationException(string.Format(
-                    "Unhandled type: {0}", model.GetType()));
-            }
-        }
-
-        /// <summary>
-        /// Deletes <see cref="AnnotationModel"/> object from annotation collection of the workspace.
-        /// </summary>
-        /// <param name="model"><see cref="AnnotationModel"/> object to remove.</param>
-        public void RemoveGroup(ModelBase model)
-        {
-            var annotation = model as AnnotationModel;
-            RemoveAnnotation(annotation);
-            annotation.Dispose();
-        }
-
-        /// <summary>
-        /// Updates <see cref="ModelBase"/> object with given xml data
-        /// </summary>
-        /// <param name="modelData">Xml data to update model</param>
-        public void ReloadModel(XmlElement modelData)
-        {
-            ModelBase model = GetModelForElement(modelData);
-            model.Deserialize(modelData, SaveContext.Undo);
-        }
-
-        /// <summary>
-        /// Creates <see cref="ModelBase"/> object by given xml data and
-        /// adds it to corresponding collection of the workspace.
-        /// </summary>
-        /// <param name="modelData">Xml data to create model</param>
-        public void CreateModel(XmlElement modelData)
-        {
-            var helper = new XmlElementHelper(modelData);
-            string typeName = helper.ReadString("type", String.Empty);
-            if (string.IsNullOrEmpty(typeName))
-            {
-                // If there wasn't a "type" attribute, then we fall-back onto
-                // the name of the XmlElement itself, which is usually the type
-                // name.
-                typeName = modelData.Name;
-                if (string.IsNullOrEmpty(typeName))
-                {
-                    string guid = helper.ReadString("guid");
-                    throw new InvalidOperationException(
-                        string.Format("No type information: {0}", guid));
-                }
-            }
-
-            if (typeName.Contains("ConnectorModel"))
-            {
-                var connector = NodeGraph.LoadConnectorFromXml(modelData,
-                    Nodes.ToDictionary(node => node.GUID));
-
-                // It is possible that in some cases connector can't be created,
-                // for example, connector connects to a custom node instance
-                // whose input ports have been changed, so connector can't find
-                // its end port owner.
-                if (connector == null)
-                {
-                    var guidAttribute = modelData.Attributes["guid"];
-                    if (guidAttribute == null)
-                    {
-                        throw new InvalidOperationException("'guid' field missing from recorded model");
-                    }
-                    undoRecorder.RecordModelAsOffTrack(Guid.Parse(guidAttribute.Value));
-                }
-                else
-                {
-                    OnConnectorAdded(connector); // Update view-model and view.
-                }
-            }
-            else if (typeName.Contains("NoteModel"))
-            {
-                var noteModel = NodeGraph.LoadNoteFromXml(modelData);
-                AddNote(noteModel);
-
-                //check whether this note belongs to a group
-                foreach (var annotation in Annotations)
-                {
-                    //this note "was" in a group
-                    if (annotation.DeletedModelBases.Any(m => m.GUID == noteModel.GUID))
-                    {
-                        annotation.AddToSelectedModels(noteModel);
-                    }
-                }
-            }
-            else if (typeName.Contains("AnnotationModel"))
-            {
-                var selectedNodes = this.Nodes == null ? null : this.Nodes.Where(s => s.IsSelected);
-                var selectedNotes = this.Notes == null ? null : this.Notes.Where(s => s.IsSelected);
-
-                var annotationModel = new AnnotationModel(selectedNodes, selectedNotes);
-                annotationModel.ModelBaseRequested += annotationModel_GetModelBase;
-                annotationModel.Disposed += (_) => annotationModel.ModelBaseRequested -= annotationModel_GetModelBase;
-                annotationModel.Deserialize(modelData, SaveContext.Undo);
-                AddNewAnnotation(annotationModel);
-            }
-
-            else if (typeName.Contains("PresetModel"))
-            {
-                var preset = new PresetModel(this.Nodes);
-                preset.Deserialize(modelData, SaveContext.Undo);
-                presets.Add(preset);
-                //we raise this property change here so that this event bubbles up through
-                //the model and to the DynamoViewModel so that presets show in the UI menu if our undo/redo
-                //created the first preset
-                RaisePropertyChanged("EnablePresetOptions");
-
-            }
-            else // Other node types.
-            {
-                NodeModel nodeModel = NodeFactory.CreateNodeFromXml(modelData, SaveContext.Undo, ElementResolver);
-
-                AddAndRegisterNode(nodeModel);
-
-                //check whether this node belongs to a group
-                foreach (var annotation in Annotations)
-                {
-                    //this node "was" in a group
-                    if (annotation.DeletedModelBases.Any(m=>m.GUID == nodeModel.GUID))
-                    {
-                        annotation.AddToSelectedModels(nodeModel);
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Gets model by GUID which is contained in given Xml data.
-        /// </summary>
-        /// <param name="modelData">Xml data to find model.</param>
-        /// <returns>Found <see cref="ModelBase"/> object.</returns>
-        public ModelBase GetModelForElement(XmlElement modelData)
-        {
-            // TODO(Ben): This may or may not be true, but I guess we should be
-            // using "System.Type" (given the "type" information in "modelData"),
-            // and determine the matching category (e.g. is this a Node, or a
-            // Connector?) instead of checking in each and every collections we
-            // have in the workspace.
-            //
-            // System.Type type = System.Type.GetType(helper.ReadString("type"));
-            // if (typeof(Dynamo.Models.NodeModel).IsAssignableFrom(type))
-            //     return Nodes.First((x) => (x.GUID == modelGuid));
-
-            var helper = new XmlElementHelper(modelData);
-            Guid modelGuid = helper.ReadGuid("guid");
-
-            ModelBase foundModel = GetModelInternal(modelGuid);
-            if (null != foundModel)
-                return foundModel;
-
-            throw new ArgumentException(
-                string.Format("Unhandled model type: {0}", helper.ReadString("type", modelData.Name)));
-        }
-
-        /// <summary>
-        /// Returns model by GUID
-        /// </summary>
-        /// <param name="modelGuid">Identifier of the requested model.</param>
-        /// <returns>Found <see cref="ModelBase"/> object.</returns>
-        public ModelBase GetModelInternal(Guid modelGuid)
-        {
-            ModelBase foundModel = (Connectors.FirstOrDefault(c => c.GUID == modelGuid)
-                ?? Nodes.FirstOrDefault(node => node.GUID == modelGuid) as ModelBase)
-                ?? (Notes.FirstOrDefault(note => note.GUID == modelGuid)
-                ?? Annotations.FirstOrDefault(annotation => annotation.GUID == modelGuid) as ModelBase
-                ?? Presets.FirstOrDefault(preset => preset.GUID == modelGuid) as ModelBase);
-
-            return foundModel;
-        }
-
-        /// <summary>
-        /// Gets model list by their GUIDs
-        /// </summary>
-        /// <param name="modelGuids">Identifiers of the requested models.</param>
-        /// <returns>All found <see cref="ModelBase"/> objects.</returns>
-        private IEnumerable<ModelBase> GetModelsInternal(IEnumerable<Guid> modelGuids)
-        {
-            var foundModels = new List<ModelBase>();
-
-            foreach (var modelGuid in modelGuids)
-            {
-                var foundModel = GetModelInternal(modelGuid);
-                if (foundModel != null)
-                    foundModels.Add(foundModel);
-            }
-
-            return foundModels;
-        }
-
-        #endregion
-
-        #region Node To Code Reconnection
-
-        /// <summary>
-        /// Checks whether the given connection is inside the node to code set or outside it.
-        /// This determines if it should be redrawn(if it is external) or if it should be
-        /// deleted (if it is internal)
-        /// </summary>
-        private static bool IsInternalNodeToCodeConnection(IEnumerable<NodeModel> nodes, ConnectorModel connector)
-        {
-            return nodes.Contains(connector.Start.Owner) && nodes.Contains(connector.End.Owner);
-        }
-
-        /// <summary>
-        /// Forms new connections from the external nodes to the Node To Code Node,
-        /// based on the connectors passed as inputs.
-        /// </summary>
-        /// <param name="externalOutputConnections">List of connectors to remake, along with the port names of the new port</param>
-        /// <param name="cbn">The new Node To Code created Code Block Node</param>
-        private List<ConnectorModel> ReConnectOutputConnections(Dictionary<ConnectorModel, string> externalOutputConnections, CodeBlockNodeModel cbn)
-        {
-            List<ConnectorModel> newConnectors = new List<ConnectorModel>();
-            foreach (var kvp in externalOutputConnections)
-            {
-                var connector = kvp.Key;
-                var variableName = kvp.Value;
-
-                //Get the start and end idex for the ports for the connection
-                var portModel = cbn.OutPorts.FirstOrDefault(
-                    port => cbn.GetRawAstIdentifierForOutputIndex(port.Index).Value.Equals(variableName));
-
-                if (portModel == null)
-                    continue;
-
-                //Make the new connection and then record and add it
-                var newConnector = ConnectorModel.Make(
-                    cbn,
-                    connector.End.Owner,
-                    portModel.Index,
-                    connector.End.Index);
-
-                newConnectors.Add(newConnector);
-            }
-            return newConnectors;
-        }
-
-        /// <summary>
-        /// Forms new connections from the external nodes to the Node To Code Node,
-        /// based on the connectors passed as inputs.
-        /// </summary>
-        /// <param name="externalInputConnections">List of connectors to remake, along with the port names of the new port</param>
-        /// <param name="cbn">The new Node To Code created Code Block Node</param>
-        private List<ConnectorModel> ReConnectInputConnections(
-            Dictionary<ConnectorModel, string> externalInputConnections, CodeBlockNodeModel cbn)
-        {
-            List<ConnectorModel> newConnectors = new List<ConnectorModel>();
-
-            foreach (var kvp in externalInputConnections)
-            {
-                var connector = kvp.Key;
-                var variableName = kvp.Value;
-
-                var endPortIndex = CodeBlockNodeModel.GetInportIndex(cbn, variableName);
-                if (endPortIndex < 0)
-                    continue;
-
-                if (Connectors.Any(c => c.End == cbn.InPorts[endPortIndex]))
-                    continue;
-
-                var newConnector = ConnectorModel.Make(
-                    connector.Start.Owner,
-                    cbn,
-                    connector.Start.Index,
-                    endPortIndex);
-
-                newConnectors.Add(newConnector);
-            }
-
-            return newConnectors;
         }
 
         #endregion
