+++ src/DynamoCore/Models/DynamoModel.cs
@@ -1278,8 +1278,8 @@ namespace Dynamo.Models
             try
             {
                 // Serialize ws into string json
-                var json = Autodesk.Workspaces.Utilities.SaveWorkspaceToJson(ws, this.LibraryServices, this.EngineController,
-                    this.Scheduler, this.NodeFactory, false, false, this.CustomNodeManager);
+                var json = Autodesk.Workspaces.Utilities.SaveWorkspaceToJson(ws, LibraryServices, EngineController,
+                    Scheduler, NodeFactory, false, false, CustomNodeManager);
                 Logger.Log(String.Format(Resources.SavingInProgress, path));
                 File.WriteAllText(path, json);
             }
